Move common classes from SPI to common util module
