+++ common/util/src/main/java/org/eclipse/dataspaceconnector/common/types/Cast.java
@@ -12,13 +12,14 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.common;
+package org.eclipse.dataspaceconnector.common.types;
 
 /**
  * Utility for performing typesafe casts.
  */
 public class Cast {
 
+    @SuppressWarnings("unchecked")
     public static <T> T cast(Object type) {
         return (T) type;
     }
