+++ extensions/atlas/src/test/java/org/eclipse/dataspaceconnector/catalog/atlas/metadata/AtlasApiTest.java
@@ -34,7 +34,7 @@ import java.util.stream.Stream;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.eclipse.dataspaceconnector.catalog.atlas.dto.Functions.*;
-import static org.eclipse.dataspaceconnector.common.ConfigurationFunctions.propOrEnv;
+import static org.eclipse.dataspaceconnector.common.configuration.ConfigurationFunctions.propOrEnv;
 
 @IntegrationTest
 public class AtlasApiTest {
