+++ extensions/filesystem/configuration-fs/build.gradle.kts
@@ -19,6 +19,8 @@ plugins {
 
 dependencies {
     api(project(":spi"))
+    implementation(project(":common:util"))
+
 }
 
 publishing {
