+++ extensions/filesystem/vault-fs/build.gradle.kts
@@ -18,6 +18,7 @@ plugins {
 
 dependencies {
     api(project(":spi"))
+    implementation(project(":common:util"))
 }
 
 publishing {
