+++ extensions/filesystem/vault-fs/src/main/java/org/eclipse/dataspaceconnector/security/fs/FsConfiguration.java
@@ -16,7 +16,7 @@ package org.eclipse.dataspaceconnector.security.fs;
 
 import org.eclipse.dataspaceconnector.spi.EdcSetting;
 
-import static org.eclipse.dataspaceconnector.common.ConfigurationFunctions.propOrEnv;
+import static org.eclipse.dataspaceconnector.common.configuration.ConfigurationFunctions.propOrEnv;
 
 public final class FsConfiguration {
