+++ launchers/junit/build.gradle.kts
@@ -24,6 +24,8 @@ dependencies {
 
     api(project(":core:bootstrap"))
 
+    testFixturesImplementation(project(":common:util"))
+
     testFixturesImplementation("org.junit.jupiter:junit-jupiter-api:5.5.2")
     testFixturesRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.5.2")
 }
