+++ samples/copy-with-nifi/transfer/src/test/java/org/eclipse/dataspaceconnector/transfer/nifi/NifiDataFlowControllerTest.java
@@ -65,7 +65,7 @@ import java.util.UUID;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.easymock.EasyMock.*;
-import static org.eclipse.dataspaceconnector.common.ConfigurationFunctions.propOrEnv;
+import static org.eclipse.dataspaceconnector.common.configuration.ConfigurationFunctions.propOrEnv;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @IntegrationTest
